ai-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
opencv-python==4.9.0.80
Pillow==10.2.0
numpy==1.26.3
python-magic-bin==0.4.14
pydantic==2.5.3
pytesseract==0.3.10
scikit-learn==1.4.0

# CNN + ViT Fraud Detection Model
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12

